Split and move test_emails.py
